+++ example/android/app/src/main/kotlin/com/gongziming/share_bridge_example/wxapi/WXEntryActivity.kt
@@ -1,0 +1,5 @@
+package com.gongziming.share_bridge_example.wxapi
+
+import com.gongziming.share_bridge_wechat.ShareBridgeWechatEntryActivity
+
+class WXEntryActivity : ShareBridgeWechatEntryActivity()