+++ packages/share_bridge_qq/lib/share_bridge_qq_platform_interface.dart
@@ -20,6 +20,7 @@ abstract class ShareBridgeQqPlatform extends PlatformInterface {
   Future<void> initialize({
     required String appId,
     String? universalLink,
+    required bool privacyGranted,
   }) {
     throw UnimplementedError('initialize() has not been implemented.');
   }